requirements.txt
PyQt5>=5.15
pywin32>=306
psutil>=5.9
mss>=9.0
Pillow>=10.0
opencv-python>=4.8
numpy>=1.24
